requirements.txt
numpy==1.20.2
pandas==1.2.4
python-dateutil==2.8.1
pytz==2021.1
scipy==1.6.2
six==1.15.0
python-dotenv==0.17.1
pymysql==1.0.2
gunicorn